feat(core): v2 live end-to-end suite and architecture docs (G3 M20)
The live suite found and fixed: named-style re-apply resets text styles;
cell paragraphs reject pageBreakBefore; column widths need widthType;
unbulleted nested items keep 36pt/level of indent that bulleting adds to
nesting; mixed-kind nesting is unrealizable (nested items take their
list's kind, with a note); rich links need Drive URLs; exact ISO page
sizes; float32 color comparison; dangling heading links; re-run tab
visibility after the create phase.

+++ src/core/model/styleValues.ts
@@ -202,7 +202,8 @@ const MONOSPACE_FONTS: ReadonlySet<string> = new Set([
 
 /** Canonicalizes an rgb channel triple, defaulting any missing channel to 0. */
 function rgbCanonical(rgb: RawColor): { blue: number; green: number; red: number } {
-  return { blue: rgb.blue ?? 0, green: rgb.green ?? 0, red: rgb.red ?? 0 };
+  // The API stores channels as float32 (17/255 reads back as 0.06666667), so compare at that precision.
+  return { blue: Math.fround(rgb.blue ?? 0), green: Math.fround(rgb.green ?? 0), red: Math.fround(rgb.red ?? 0) };
 }
 
 /** Converts an rgb 0-1 channel triple (missing channels default to 0) into `#RRGGBB`. */
